+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 Все крупные изменения в проекте будут отражены в этом файле.
 
-## 4.16-release (31.08.2023)
+## 4.16-release (28.09.2023)
+* Добавлен обязательный заголовок сжатия `Accept-Encoding:gzip` для всех запросов
+* В README обновлен домен remap-12 на `api.moysklad.ru`
+
+## 4.15-release (31.08.2023)
 * Добавлено поле `allowDeleteReceiptPositions` в `RetailStore`
 
-## 4.15-release (10.08.2023)
+## 4.14-release (10.08.2023)
 * Добавлены поля `sex`, `birthDate` в `Counterparty`
 
 ## 4.11-release (14.09.2022)